Fix for bug #292
Modified the workaround written for #292 to instead "repairing" the data bindings to have a separate
constructor that doesn't use DataDictionary to retrieve field type and field name

The main problem was that both PutOrCall and OrderCapacity are present in some FIX versions and not in others, so the data dictionary was not finding them.

now we just manually specify the field type/name instead of relying on data dictionary to look it up.




git-svn-id: http://code.marketcetera.org/root/trunk@6714 501a4a96-914b-0410-8454-f5988a4ca596