Reviewed as ENG-271, changes as follows:
This change represents the modifications to Strategy necessary for Platform As a Service.  The specific changes are below.  All unit tests pass.  Current functionality is 
maintained with the exception of notifications.  Notifications created in strategy scripts will not be displayed in Photon until corresponding changes are made in Photon.  To 
test the output messages, I ran all unit tests with logger set to debug and captured the output.  I examined the output for each message.

1) Change Strategy Module signature:
 a) One URN for all outputs
 b) Remove Classpath
 c) Add boolean to indicate whether to send orders to the ORS
2) Make EventBase subclasses serializable with UID
3) Remove unused ExecutionReport
4) Add LogEvent to represent log output from strategies
5) Make MarketDataRequest serializable
6) Create new strategy output streams
 a) Notifications (and remove current plumbing to notification manager)
 b) Log
7) Create new Strategy API for logging
8) Reroute all Strategy Log4J outputs to "strategy.messages" to the output URN



git-svn-id: http://code.marketcetera.org/root/trunk@10310 501a4a96-914b-0410-8454-f5988a4ca596