Fix for EG-658, memory leak in DOB Photon display.
 The problem was that Bogus and Activ did not adhere to the same contract with respect to the pattern of events emitted.  This is due to a design oversight, to wit, a 
failure to recognize that such a thing was required.  While each was externally consistent, their lack of mutual consistency meant that the data for Bogus was not 
being displayed properly in Photon.  The fixes included correct this problem by making Bogus behave the same way as Activ.

Reviewed as ENG-415



git-svn-id: http://code.marketcetera.org/root/trunk@10567 501a4a96-914b-0410-8454-f5988a4ca596