Fix for EG-401, reviewed as ENG-428:
This fix is to allow the execution of Strategy unit tests on Windows without (most) path restrictions. The fix is to, for unit tests, piggy-back on Maven's manifest 
jar construction to get the proper classpath for Java strategy compilation. This fix is a bit hacky, but is invoked only for unit tests. As a side-effect, there will 
exist an undocumented method to augment the Java strategy classpath if necessary. I can't see a reason for doing this off-hand, but the possibility will exist.

I ran unit tests on Windows and Linux successfully. I also installed a 1.6 server on Windows and swapped out the strategy jar to make sure it works in production. The 
fix shouldn't apply to production, just wanted to make sure nothing broke.

There is a bonus fix included to a client test that allows a more generous timeout. I encountered this problem on my Windows VM. T2 approved (suggested, in fact) the 
fix. Lastly, I found a problem if the JVM used to execute strategy is a JRE (instead of a JDK). In that case, there is no compiler included in the installation and 
the Java strategies will all fail. I just added a message to catch this particular case.



git-svn-id: http://code.marketcetera.org/root/trunk@10590 501a4a96-914b-0410-8454-f5988a4ca596