Fix for EG-727, Strategy Parameters Cannot Contain ":" or "="
Strategy parameters are set by creating a Properties map from a flat string.  The string expects key/value pairs to be separated by "=" and delimited by ":", as in: 
key1=value1:key2=value2.  The problem comes when the key or value tokens actually contain ":" or "=".  This is not so far-fetched, as a reasonable parameter value may 
be a Windows path like "c:\some-path".  The parsing code took: file=c:\some-path and yielded: [file=c] (the rest was discarded) as a malformed key/value pair.

The fix is to add logic that allows ":" and "=" to be escaped using the "\" char.  To repeat the example above, then, file=c\:\some-path and yields: 
[file=c:\some-path].  Unit tests and manual testing pass.

Reviewed by T2 as ENG-463



git-svn-id: http://code.marketcetera.org/root/trunk@10656 501a4a96-914b-0410-8454-f5988a4ca596